+++ src/ws/types.ts
@@ -118,46 +118,48 @@ export type WSParseHandler<Route extends RouteSchema, Context = {}> = (
 	message: unknown
 ) => MaybePromise<Route['body'] | void | undefined>
 
-export type AnyWSLocalHook = WSLocalHook<any, any, any>
+export type AnyWSLocalHook = WSLocalHook<any, any, any, any>
 
 export type WSLocalHook<
 	LocalSchema extends InputSchema,
 	Schema extends RouteSchema,
-	Singleton extends SingletonBase
-> = (LocalSchema extends any ? LocalSchema : Prettify<LocalSchema>) & {
-	detail?: DocumentDecoration
-	/**
-	 * Headers to register to websocket before `upgrade`
-	 */
-	upgrade?: Record<string, unknown> | ((context: Context) => unknown)
-	parse?: MaybeArray<WSParseHandler<Schema>>
+	Singleton extends SingletonBase,
+	Macro extends MetadataBase['macro']
+> = Prettify<Macro> &
+	(LocalSchema extends any ? LocalSchema : Prettify<LocalSchema>) & {
+		detail?: DocumentDecoration
+		/**
+		 * Headers to register to websocket before `upgrade`
+		 */
+		upgrade?: Record<string, unknown> | ((context: Context) => unknown)
+		parse?: MaybeArray<WSParseHandler<Schema>>
 
-	/**
-	 * Transform context's value
-	 */
-	transform?: MaybeArray<TransformHandler<Schema, Singleton>>
-	/**
-	 * Execute before main handler
-	 */
-	beforeHandle?: MaybeArray<OptionalHandler<Schema, Singleton>>
-	/**
-	 * Execute after main handler
-	 */
-	afterHandle?: MaybeArray<OptionalHandler<Schema, Singleton>>
-	/**
-	 * Execute after main handler
-	 */
-	mapResponse?: MaybeArray<MapResponse<Schema, Singleton>>
-	/**
-	 * Execute after response is sent
-	 */
-	afterResponse?: MaybeArray<AfterResponseHandler<Schema, Singleton>>
-	/**
-	 * Catch error
-	 */
-	error?: MaybeArray<ErrorHandler<{}, Schema, Singleton>>
-	tags?: DocumentDecoration['tags']
-} & TypedWebSocketHandler<
+		/**
+		 * Transform context's value
+		 */
+		transform?: MaybeArray<TransformHandler<Schema, Singleton>>
+		/**
+		 * Execute before main handler
+		 */
+		beforeHandle?: MaybeArray<OptionalHandler<Schema, Singleton>>
+		/**
+		 * Execute after main handler
+		 */
+		afterHandle?: MaybeArray<OptionalHandler<Schema, Singleton>>
+		/**
+		 * Execute after main handler
+		 */
+		mapResponse?: MaybeArray<MapResponse<Schema, Singleton>>
+		/**
+		 * Execute after response is sent
+		 */
+		afterResponse?: MaybeArray<AfterResponseHandler<Schema, Singleton>>
+		/**
+		 * Catch error
+		 */
+		error?: MaybeArray<ErrorHandler<{}, Schema, Singleton>>
+		tags?: DocumentDecoration['tags']
+	} & TypedWebSocketHandler<
 		Omit<Context<Schema, Singleton>, 'body'> & {
 			body: never
 		},